CosineGreedyNumba -> CosineGreedy and former CosineGreedy -> CosineGreedyVectorial
