+++ matchms/similarity/CosineGreedy.py
@@ -1,17 +1,18 @@
 from typing import Tuple
 import numpy
+from matchms.similarity.collect_peak_pairs import collect_peak_pairs
 from matchms.typing import SpectrumType
 
 
 class CosineGreedy:
-    """Factory to calculate 'greedy cosine score' between mass spectra.
+    """Calculate 'cosine similarity score' between two spectra.
 
     The cosine score aims at quantifying the similarity between two mass spectra.
     The score is calculated by finding best possible matches between peaks
     of two spectra. Two peaks are considered a potential match if their
     m/z ratios lie within the given 'tolerance'.
-    The underlying peak assignment problem is here solved in a 'greedy' way.
-    This can perform notably faster, but does occasionally deviate slightly from
+    The underlying peak assignment problem is here sovled in a 'greedy' way.
+    This can perform notably faster, but does occationally deviate slightly from
     a fully correct solution (as with the Hungarian algorithm). In practice this
     will rarely affect similarity scores notably, in particular for smaller
     tolerances.
@@ -43,81 +44,56 @@ class CosineGreedy:
         Cosine score is 0.52 with 1 matched peaks
 
     """
-    def __init__(self, tolerance: float = 0.3):
+    def __init__(self, tolerance=0.1):
         """
-
-        Parameters
-        ----------
-        tolerance
-            Peaks will be considered a match when <= tolerance apart.
+        Args:
+        ----
+        tolerance: float
+            Peaks will be considered a match when <= tolerance appart. Default is 0.1.
         """
         self.tolerance = tolerance
 
-    def __call__(self, spectrum: SpectrumType, reference_spectrum: SpectrumType) -> Tuple[float, int]:
-        """Calculate 'greedy cosine score' between mass spectra.
-
+    def __call__(self, spectrum1: SpectrumType, spectrum2: SpectrumType) -> Tuple[float, int]:
+        """Calculate cosine score between two spectra.
         Args:
-        -----
-        spectrum
-            First spectrum
-        reference_spectrum
-            Second spectrum
-
-        Returns:
-        --------
-
-        Tuple with cosine score and number of matched peaks.
+        ----
+        spectrum1: SpectrumType
+            Input spectrum 1.
+        spectrum2: SpectrumType
+            Input spectrum 2.
         """
-        def calc_mz_distance():
-            mz_row_vector = spectrum.peaks.mz
-            mz_col_vector = numpy.reshape(reference_spectrum.peaks.mz, (n_rows, 1))
+        def get_peaks_arrays():
+            """Get peaks mz and intensities as numpy array."""
+            spec1 = numpy.vstack((spectrum1.peaks.mz, spectrum1.peaks.intensities)).T
+            spec2 = numpy.vstack((spectrum2.peaks.mz, spectrum2.peaks.intensities)).T
+            assert max(spec1[:, 1]) <= 1, ("Input spectrum1 is not normalized. ",
+                                           "Apply 'normalize_intensities' filter first.")
+            assert max(spec2[:, 1]) <= 1, ("Input spectrum2 is not normalized. ",
+                                           "Apply 'normalize_intensities' filter first.")
+            return spec1, spec2
 
-            mz1 = numpy.tile(mz_row_vector, (n_rows, 1))
-            mz2 = numpy.tile(mz_col_vector, (1, n_cols))
-
-            return mz1 - mz2
-
-        def calc_intensities_product():
-            intensities_row_vector = spectrum.peaks.intensities
-            intensities_col_vector = numpy.reshape(reference_spectrum.peaks.intensities, (n_rows, 1))
-
-            intensities1 = numpy.tile(intensities_row_vector, (n_rows, 1))
-            intensities2 = numpy.tile(intensities_col_vector, (1, n_cols))
-
-            return intensities1 * intensities2
-
-        def calc_intensities_product_within_tolerance():
-
-            mz_distance = calc_mz_distance()
-            intensities_product = calc_intensities_product()
-
-            within_tolerance = numpy.absolute(mz_distance) <= self.tolerance
-
-            return numpy.where(within_tolerance, intensities_product, numpy.zeros_like(intensities_product))
+        def get_matching_pairs():
+            """Get pairs of peaks that match within the given tolerance."""
+            matching_pairs = collect_peak_pairs(spec1, spec2, self.tolerance, shift=0.0)
+            matching_pairs = sorted(matching_pairs, key=lambda x: x[2], reverse=True)
+            return matching_pairs
 
         def calc_score():
-            r_unordered, c_unordered = intensities_product_within_tolerance.nonzero()
-            v_unordered = intensities_product_within_tolerance[r_unordered, c_unordered]
-            sortorder = numpy.argsort(v_unordered)[::-1]
-            r_sorted = r_unordered[sortorder]
-            c_sorted = c_unordered[sortorder]
+            """Calculate cosine similarity score."""
+            used1 = set()
+            used2 = set()
+            score = 0.0
+            used_matches = []
+            for match in matching_pairs:
+                if not match[0] in used1 and not match[1] in used2:
+                    score += match[2]
+                    used1.add(match[0])  # Every peak can only be paired once
+                    used2.add(match[1])  # Every peak can only be paired once
+                    used_matches.append(match)
+            # Normalize score:
+            score = score/max(numpy.sum(spec1[:, 1]**2), numpy.sum(spec2[:, 1]**2))
+            return score, len(used_matches)
 
-            score = 0
-            n_matches = 0
-            for r, c in zip(r_sorted, c_sorted):
-                if intensities_product_within_tolerance[r, c] > 0:
-                    score += intensities_product_within_tolerance[r, c]
-                    n_matches += 1
-                    intensities_product_within_tolerance[r, :] = 0
-                    intensities_product_within_tolerance[:, c] = 0
-            return score / max(sum(squared1), sum(squared2)), n_matches
-
-        n_rows = reference_spectrum.peaks.mz.size
-        n_cols = spectrum.peaks.mz.size
-
-        intensities_product_within_tolerance = calc_intensities_product_within_tolerance()
-
-        squared1 = numpy.power(spectrum.peaks.intensities, 2)
-        squared2 = numpy.power(reference_spectrum.peaks.intensities, 2)
-
+        spec1, spec2 = get_peaks_arrays()
+        matching_pairs = get_matching_pairs()
         return calc_score()
